move macros to new file so lib.rs will be reduced bug fix of wolf variant
